Adds check for plugins to Bib::Vagrant and moves config code to its own file, plus update README. easybib/issues#5384
